@@ -20,7 +20,7 @@
     ],
     "minimum-stability": "stable",
     "require": {
-        "php": ">=5.3.9",
+        "php": ">=5.4",
         "symfony/http-kernel": "~2.8 || ~3.0",
         "symfony/dependency-injection": "~2.8 || ~3.0",
         "symfony/http-foundation": "~2.8 || ~3.0",
