@@ -8,11 +8,11 @@ class DictionarySpec extends ObjectBehavior
 {
     function let()
     {
-        $this->beConstructedWith('foo', array(
+        $this->beConstructedWith('foo', [
             'foo' => 0,
             'bar' => 1,
             'baz' => 2,
-        ));
+        ]);
     }
 
     function it_is_initializable()
@@ -22,11 +22,11 @@ class DictionarySpec extends ObjectBehavior
 
     function its_getvalues_should_return_dictionary_values()
     {
-        $this->getValues()->shouldReturn(array(
+        $this->getValues()->shouldReturn([
             'foo' => 0,
             'bar' => 1,
             'baz' => 2,
-        ));
+        ]);
     }
 
     function its_getname_should_return_dictionary_name()
